Move utility components out of components/ dir
