+++ stories/1-unread-messages.md
@@ -18,6 +18,6 @@ Field `messages` on `Member` contains a list of messages, each with a boolean `r
 
 ## Feature Requirements
 
-1. An `<UnreadMessageTag />` should display to the right of each member name. This component exists already but is not rendered anywhere.
+1. An `<UnreadMessageTag />` should display to the right of each member name. This component is already written, but is not yet rendered anywhere.
 2. The tag should indicate the number of unread messages.
 3. Users with zero unread messages should not have a tag.
